requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Dance==7.1.0
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
Flask-WTF==1.2.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
oauthlib==3.3.0
packaging==25.0
python-dotenv==1.1.0
requests==2.32.4
requests-oauthlib==2.0.0
urllib3==2.5.0
URLObject==2.4.3
Werkzeug==3.1.3
WTForms==3.2.1
